refactor: extract tensor/array utilities into tensorutils module
Move the stateless create_bounded_tensor and ensure_array (formerly
Model._ensure_array) helpers into a new src/pyhs3/tensorutils.py module.
Neither function accesses Model state, so they were awkwardly placed;
tensorutils gives them a single-responsibility home.

Reduces model.py from 1014 to 957 lines, clearing pylint max-module-lines
(1000) notice. Pylint score remains 10.00/10.
